Rename to `type_assertions.rb`
