+++ lib/rbs/unit_test.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 require "rbs/test"
 require "rbs/unit_test/spy"
-require "rbs/unit_test/type_assertion"
+require "rbs/unit_test/type_assertions"
